Refactor shared out of curve.rs
