+++ token-swap/program/src/curve/mod.rs
@@ -1,0 +1,2 @@
+//! ...
+pub mod shared;
