+++ token-swap/program/src/instruction.rs
@@ -2,7 +2,7 @@
 
 #![allow(clippy::too_many_arguments)]
 
-use crate::curve::SwapCurve;
+use crate::curve::shared::SwapCurve;
 use crate::error::SwapError;
 use solana_program::{
     instruction::{AccountMeta, Instruction},
@@ -381,7 +381,7 @@ pub fn unpack<T>(input: &[u8]) -> Result<&T, ProgramError> {
 mod tests {
     use super::*;
 
-    use crate::curve::{CurveType, FlatCurve};
+    use crate::curve::shared::{CurveType, FlatCurve};
 
     #[test]
     fn test_instruction_packing() {
